Fix SBT runners to escape quotes nicely.
This is an impossible problem, in general. See
https://groups.google.com/d/topic/simple-build-tool/el5c3oaSgsw/discussion

Essentially, SBT made the decision to parse the command-line on its own.
This means we have to pass the entire command to SBT as a single
argument. _That_ means we have to escape arguments ourselves, in Bash.
_That_ means we have to depend upon Bash instead of just a Bourne shell.
And after all that, I don't think backslashes are handled correctly. (I'm
not sure, though.)

[refs #66799388] to pass an SBT argument like '; project runner; test'