Fiddle with position of document view's ul.actions
This is a partial fix, but it [finishes #69527356].

At issue: if text goes all the way over to the right of the document
view, the ul.actions covers it up. This manifests as a splotch of
background color over the bottom half of the line of text.

This fix assumes the last line of the document-list-cursor is always the
clustering keywords. It lowers the top of the ul.actions a bit, so it
always lines up with the last line of text. At worst, that line is cut
off a bit (with a pleasant gradient). But that line is often so short it
doesn't get cut off at all.