Prevent race when submitting quick jobs.
Before, if a job finished at the same time as the user browsed to
/documentsets (which is where the job-submit page directs them), the
worker could delete the job in between the server's queries for
"document sets" and "jobs". (This is due to our silly database design in
which document sets which are _missing_ jobs are normal.) The result:
the server would see nothing for the split second of the request.

This should solve intermittent Travis errors.