:s/encodedFilename/filename
It isn't encoded any more.

I mean, in the grander scheme of things it is, because it's a Java
String and all Java Strings are UTF-16 encoded. But it isn't any _more_
encoded than any other string we use. So the choice was either to prefix
every string with the word "encoded" or to remove the word "encoded"
here.

[refs #81786576]