Make PDF documents viewable again
[finishes #88895202]

I introduced this regression in a bout of Squeryl hunting. It turns out
OverviewDocument.url was designed to create a viewable URL, whereas
Document.url was meant to be a database column. The latter is clearly
what we should have done all along. I added a viewUrl method to derive a
viewable Url.