Put DocumentDisplayPreferences in the Show app
[finishes #90522844]

This isn't the dream: keywords don't have spaces, so they can't wrap;
and even if they could, the float:right and float:left force keywords
and preferences to divvy up the horizontal space instead of making
keywords flow around prefs. One viable workaround is to revamp keywords
and make them and prefs fit in a big flexbox; that's a story for another
day though.