Create a new 'Browser' integration-test framework
This uses selenium-webdriver instead of wd. Why? Because wd development
seems to be dead. That said, selenium-webdriver uses a convoluted
promise library to emulate synchronous execution. It seems the
developers did test-driven development rather than build a logical API;
the result is bug after bug after bug in real-world code.

If this doesn't work, we'll need to switch to Scala. Node is bad for
this because we want tests to be synchronous.

The problem: Jenkins was failing on Recluster, but it was passing on
dev machines. That's the one spec I rewrote.