Show "Log Back In" Modal when user's session has expired
Notes:
- This shows whenever the user triggers an action on the page and their
session has expired in the background. The modal won’t show, though, on
the user clicking a document in the document list from the main Show
view, unfortunately, (at least if the new doc requested is a pdf)
because the request for that new doc happens in a separate worker, and
jquery isn't aware of it.

- This commit also renames the password.secure to
authentication.password.secure to group it with the new