Kill the digest stuff. It doesn't work.
I tried changing this for Play 2.4, because Play's docs say to do it.
Really, this was a Play 2.3 feature. We can't use it because when we
specify RjsKeys.modules that kills sbt-web's pipeline. Part of me wants
to file this bug with sbt-rjs; part of me thinks this is a good excuse
to ditch the asset pipeline entirely, in favor of nodejs+gulp.