Upgrade AsyncHttpClient. Nix some Java-isms.
Basically, I've removed an Akka actor or two that was effectively
emulating AsyncHttpClient or ExecutionContext. I also nixed the
distinction between HTTP error and non-200 HTTP response, because the
end user doesn't care.

During my travails, I noticed that this whole system is, erm, terrible.
Hopefully today's work is a first step towards revamping it and moving
it to DocumentSetWorker.

[finishes #103497960]

Squashed commit of the following:

commit d29e47bd68657a39d1c974e87ca2a34844ee8c33
Author: Adam Hooper <adam@adamhooper.com>
Date:   Wed Sep 16 17:42:52 2015 -0400

    Fix everything up

commit d9ed4ccb63034ceb213d6c33e059a04f30d263d5
Author: Adam Hooper <adam@adamhooper.com>
Date:   Wed Sep 16 13:17:47 2015 -0400

    Upgrade AsyncHttpClient. Nix Java-isms.

    Tests don't compile because of a Scala compiler bug. Rather than delve
    too deeply, I'll just nix the use of Akka in DocumentRetriever.