Use Enumerator in export, so it really streams.
This nixes the Content-Length header, which means the browser won't show
a progress bar any more.

There's a case to be made for "show progress bar if we can calculate the
length within 1s, otherwise don't"; but this is a surprising amount of
development effort, and it would only show progress bars for small
document sets -- the ones where the progress bar isn't even necessary.

The only other good approach is to create the entire file, *then* start
streaming it. Google Docs does this when you export a zipfile of a
folder. It means more work and more server resources -- and it actually
delivers files more *slowly* -- so I'll wait for a user request before
implementing it.

[finishes #103411238]