Find monotonic ranges when deleting documents.
We've been running into a problem on prod, reindexing a 2M-doc docset
filled with tiny documents. The problem is addDocuments() does an
implicit delete of all the document IDs. We're missing an optimization
(that's this commit!) that reduces the query complexity. Since each
Lucene commit involves so many documents (because they're tiny), the
query Lucene runs to find deleted documents is enormous (because there
are so many IDs, and before these commits each ID is another Query).

Now, we find monotonically-increasing NumericRanges of document IDs.
We assume that if Postgres doesn't give us documents in sequential ID
order, it will at least give us "clusters" of sequential-ID documents.
Each sequence becomes one Query. Assuming a small number of Queries, the
commits and merges won't have much math to do.

If this doesn't work on prod, we'll have to differentiate between
addDocuments() and updateDocument().