Render invalid-search-query errors without crash
The crash was because Oboe insists on parsing the server error the same
way it parses the server success response -- in this case, outputting
documents. But the error response is different, and it shouldn't be
rendered as though it's successful.

In the process, I noticed the DocumentListTitleView nested an <h3>
within an <h3>. I fixed that -- it changed the style a bit.