Handle soffice.bin retval 81.
It's an odd one: soffice.bin returns 81 if the profile directory
doesn't exist. But it creates the profile directory. So we just
have to re-run it.

We `rm -rf` the profile directory first, so we always know what
retval to expect.

This fixes a bug on production: after each deploy, the first four ODT
documents aren't converted.

Also: now we hard-code the soffice.bin path. Previously we made it
configurable and we defaulted to `soffice`, which behaves differently.
(This is why all our tests passed.) We cannot use `soffice` because
we kill it on timeout.