Clean docset command queue on FileGroup complete
DocumentSet commands are another area we ought to convert to streams.
But in the short run, a bit of a hack does the trick.

Also, I removed the bit where FileGroupSource reads pending tasks from
the database. It was spurious -- we already do the same in
WorkerActorEnvironment -- but more than that, it was misplaced: if we
resume within the Source, we don't have the context to know we need
to clear the DocumentSet command queue on completion.