Nix Decider output buffer
It conflicts with the "preferred" merge. If 100 documents come in and
they're all supposed to recurse and the buffer is 100, then all 100
will be Decided and waiting for a step; the recursing outputs will then
arrive, but they'll be queued behind those initial 100.

The buffer wasn't providing any value, either.