docker-run.sh: use `docker exec` when we can
This means we run web, worker, and (eventually) converters on the same
Docker container. This isn't ideal; but it unifies the two common cases:

1. `./dev` runs both Web and Worker -- hence, Worker is on localhost
2. Developer runs `./sbt web/run` in one tab and `./sbt worker/~reStart`
   in another -- hence, before this commit, Worker is in a separate
   container.

Ideally they'd always run in separate containers and we'd use DNS to
connect them. But `./sbt` is so slow and memory-intensive that we
optimize by only running it once in the common, `./dev` case. That means
we need to configure the `worker` and `web` to connect to one another
via `localhost`.

This commit makes `localhost` work in case 2.