Implement HttpTaskServer
There's an icky race: the last element of any POST is only _queued_ when
the request goes through. That's far more likely to hinder testing than
actual workers -- and indeed, it did. A few judicious Thread.sleeps and
a MergeHub buffer size of 1 got me through 20 repeated tests without
erroring because of a race.