Start refactor to consume before HTTP response
There's a design flaw: HttpTaskHandler produces StepOutputFragments
before their bytes have been consumed. (e.g., Blob contains a Source,
and we don't know when that Source is emptied). I had hoped akka-http's
Multipart.FormData Source would consume everything before complete, but
it doesn't seem to -- either that, or there's some other misbehavior I'm
not seeing.

Regardless of the exact bug, I want to simplify this Source-of-Sources
nonsense, converting to a normal Source[StepOutputElement] as soon as
possible. When HttpTaskHandler outputs StepOutputElement, that will
either fix the bug entirely, or it will make it easier to squash.