Bump RAM for pdfocr
We had problems in production: tesseract exits with error 137. That's
what we'd see if we were out of memory. I tested on the problem file
(which a user sent us) and its max RSS (according to my quick `top`
refreshes) was 170MB. So I don't know whether this will fix anything.

More promising: we'll bump deps on pdfocr, which should fix some
bugs. And we'll use tini to avoid zombie processes.