docker/unit-test: fix order of docker commands
Previously, `docker network rm` output was being ignored (because we
expected failure in the common case). But something -- probably a
Docker upgrade upstream -- suddenly made us hide an error message.

The error: we tried to delete the network before the containers that
relied on it.

The fix: 1) only cleanup containers+networks that actually exist;
2) fix the order of removals.