Upgrade to Amazon EKS
We deleted a hell of a lot more than we added.

I moved the database to Amazon RDS, so we don't need to manage its
state. The only stateful volume is the search index.

TODO configure logging.
TODO split out SearchIndex (make it a gRPC server) to be a StatefulSet.
TODO move converters to Fargate Spot.
TODO split out the Tree worker and move it to Fargate Spot.